requirements.txt
matplotlib
pandas
numpy
yfinance
scipy
flask
kiteconnect
python-dotenv
